feat: Support asterisk (*) version specifiers in package dependencies
This feature allows to declare package dependencies and versions on a project level and use an asterisk version specifier on the package level.
When building a package the asterisk specifier will then be replaced with the version specified s part of the package build process with the version specifier used for the same dependency on the project level. While each package must still declare its dependency the feature simplifies version management if multiple individual packages share a common dependency and dependency version.